ajout fichier print, leak free
