+++ makefile
@@ -6,7 +6,7 @@
 #    By: emauduit <emauduit@student.42.fr>          +#+  +:+       +#+         #
 #                                                 +#+#+#+#+#+   +#+            #
 #    Created: 2024/01/26 16:12:37 by azbk              #+#    #+#              #
-#    Updated: 2024/02/26 16:18:15 by emauduit         ###   ########.fr        #
+#    Updated: 2024/02/26 16:46:31 by emauduit         ###   ########.fr        #
 #                                                                              #
 # **************************************************************************** #
 
@@ -31,6 +31,7 @@ SRC_DIR = ./src
 ################################################################################
 
 SRC :=	main.c \
+		print_a_suppr.c \
 		\
 		builtins/cd/cd.c \
 		builtins/echo/echo.c \
